--- a/cpp/src/decisiontree/batched-levelalgo/split.cuh
+++ b/cpp/src/decisiontree/batched-levelalgo/split.cuh
@@ -74,10 +74,16 @@ struct Split {
    * @brief updates the current split if the input gain is better
    */
   DI void update(const SplitT& other) {
-    if (other.best_metric_val == best_metric_val) {
-      if (other.colid < colid) *this = other;
-    } else if (other.best_metric_val > best_metric_val) {
+    if (other.best_metric_val > best_metric_val) {
       *this = other;
+    } else if (other.best_metric_val == best_metric_val) {
+       if (other.colid > colid) {
+        *this = other;
+      } else if (other.colid == colid) {
+        if (other.quesval > quesval) {
+          *this = other;
+        }
+      }
     }
   }