@vinaydes 2021-04-22
I tested and found that the RF classification performance is unchanged due to this fix.
**Before fix:**
```
./sg_benchmark --benchmark_filter=RFClassifier*
2021-04-22 14:11:39
Running ./sg_benchmark
Run on (16 X 4300 MHz CPU s)
CPU Caches:
  L1 Data 32K (x8)
  L1 Instruction 32K (x8)
  L2 Unified 1024K (x8)
  L3 Unified 11264K (x1)
Load Average: 5.15, 5.75, 7.18
***WARNING*** CPU scaling is enabled, the benchmark real time measurements may be noisy and will incur extra overhead.
-----------------------------------------------------------------------------------
Benchmark                                         Time             CPU   Iterations
-----------------------------------------------------------------------------------
RFClassifier<float>/blobs/0/manual_time        1093 ms          771 ms            1
RFClassifier<float>/blobs/1/manual_time        2493 ms         2310 ms            1
RFClassifier<float>/blobs/2/manual_time        3666 ms         3665 ms            1
RFClassifier<float>/blobs/3/manual_time        5928 ms         5836 ms            1
RFClassifier<float>/blobs/4/manual_time       23189 ms        23116 ms            1
RFClassifier<float>/blobs/5/manual_time       41318 ms        41034 ms            1
RFClassifier<double>/blobs/0/manual_time       1259 ms          949 ms            1
RFClassifier<double>/blobs/1/manual_time       2598 ms         2420 ms            1
RFClassifier<double>/blobs/2/manual_time       4111 ms         4104 ms            1
RFClassifier<double>/blobs/3/manual_time       6504 ms         6366 ms            1
```
**After fix:**
```
./sg_benchmark --benchmark_filter=RFClassifier*
2021-04-22 14:19:26
Running ./sg_benchmark
Run on (16 X 4300 MHz CPU s)
CPU Caches:
  L1 Data 32K (x8)
  L1 Instruction 32K (x8)
  L2 Unified 1024K (x8)
  L3 Unified 11264K (x1)
Load Average: 7.31, 6.58, 7.12
***WARNING*** CPU scaling is enabled, the benchmark real time measurements may be noisy and will incur extra overhead.
-----------------------------------------------------------------------------------
Benchmark                                         Time             CPU   Iterations
-----------------------------------------------------------------------------------
RFClassifier<float>/blobs/0/manual_time        1088 ms         1052 ms            1
RFClassifier<float>/blobs/1/manual_time        2486 ms         2289 ms            1
RFClassifier<float>/blobs/2/manual_time        3672 ms         3638 ms            1
RFClassifier<float>/blobs/3/manual_time        5935 ms         5772 ms            1
RFClassifier<float>/blobs/4/manual_time       23245 ms        23184 ms            1
RFClassifier<float>/blobs/5/manual_time       41321 ms        41267 ms            1
RFClassifier<double>/blobs/0/manual_time       1258 ms         1212 ms            1
RFClassifier<double>/blobs/1/manual_time       2592 ms         2403 ms            1
RFClassifier<double>/blobs/2/manual_time       4117 ms         4113 ms            1
RFClassifier<double>/blobs/3/manual_time       6487 ms         6345 ms            1
```